change asserted favorite merchant for customer 66
I think this might have been a case of transposing 49 to 94 somewhere along the way. I went into rails console and found customer 66. They have 9 successful invoices and 2 of those are for merchant 49. Merchant 49 is the only merchant where they have multiple successful invoices so it must be returned as the favorite. Should be able to replicate this in your console if you want to confirm on your end.