Change expected amount for merchant total by date
The total 8116.35 is the revenue total for 'merchant_id_two=3' for the date 'date_one'. This could have been a simple mistake when gathering the data for assertions. The total 4671.73 matches the total for 'merchant_id_two=3' on the date 'date_two'. You should be able to replicate this on your end to confirm.