added an 'assert_equal_to_either' method and tweaked one in list.
One in list's message, while logically true, wasn't as helpful as it could be.
I have two failing tests in the example rales_engine because my created_at is
not working properly, but I don't have time to fix it. The new assertion
method isn't ideal, but it does work.